+++ packages/clerk-js/src/core/resources/Verification.ts
@@ -1,6 +1,9 @@
 import { parseError } from '@clerk/shared/error';
 import type {
   ClerkAPIError,
+  PasskeyVerificationResource,
+  PublicKeyCredentialCreationOptionsJSON,
+  PublicKeyCredentialCreationOptionsWithoutExtensions,
   SignUpVerificationJSON,
   SignUpVerificationResource,
   SignUpVerificationsJSON,
@@ -11,6 +14,7 @@ import type {
 } from '@clerk/types';
 
 import { unixEpochToDate } from '../../utils/date';
+import { convertJSONToPublicKeyCreateOptions } from '../../utils/passkeys';
 import { BaseResource } from './internal';
 
 export class Verification extends BaseResource implements VerificationResource {
@@ -48,6 +52,27 @@ export class Verification extends BaseResource implements VerificationResource {
       this.attempts = data.attempts;
       this.expireAt = unixEpochToDate(data.expire_at);
       this.error = data.error ? parseError(data.error) : null;
+    }
+    return this;
+  }
+}
+
+export class PasskeyVerification extends Verification implements PasskeyVerificationResource {
+  publicKey: PublicKeyCredentialCreationOptionsWithoutExtensions | null = null;
+
+  constructor(data: VerificationJSON | null) {
+    super(data);
+    this.fromJSON(data);
+  }
+
+  /**
+   * Transform base64url encoded strings to ArrayBuffer
+   */
+  protected fromJSON(data: VerificationJSON | null): this {
+    if (data?.nonce) {
+      this.publicKey = convertJSONToPublicKeyCreateOptions(
+        JSON.parse(data.nonce) as PublicKeyCredentialCreationOptionsJSON,
+      );
     }
     return this;
   }