+++ packages/types/src/verification.ts
@@ -13,6 +13,14 @@ export interface VerificationResource extends ClerkResource {
   verifiedFromTheSameClient: () => boolean;
 }
 
+export type PublicKeyCredentialCreationOptionsWithoutExtensions = Omit<
+  Required<PublicKeyCredentialCreationOptions>,
+  'extensions'
+>;
+export interface PasskeyVerificationResource extends VerificationResource {
+  publicKey: PublicKeyCredentialCreationOptionsWithoutExtensions | null;
+}
+
 export type VerificationStatus = 'unverified' | 'verified' | 'transferable' | 'failed' | 'expired';
 
 export interface CodeVerificationAttemptParam {