+++ packages/chrome-extension/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 1.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies [[`8350f73a6`](https://github.com/clerk/javascript/commit/8350f73a67f8980be78e3bd3343e772f5653d718)]:
+  - @clerk/shared@2.0.0-beta.19
+  - @clerk/clerk-js@5.0.0-beta.32
+  - @clerk/clerk-react@5.0.0-beta.29
+
+## 1.0.0-beta.31
+
+### Patch Changes
+
+- Updated dependencies [[`fafa76fb6`](https://github.com/clerk/javascript/commit/fafa76fb66585b5836cc79985f8bdf1d1b4dca97)]:
+  - @clerk/clerk-js@5.0.0-beta.31
+  - @clerk/clerk-react@5.0.0-beta.28
+
 ## 1.0.0-beta.30
 
 ### Patch Changes