+++ packages/gatsby-plugin-clerk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 5.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @clerk/backend@1.0.0-beta.27
+  - @clerk/clerk-react@5.0.0-beta.29
+  - @clerk/clerk-sdk-node@5.0.0-beta.28
+
+## 5.0.0-beta.31
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @clerk/clerk-react@5.0.0-beta.28
+
 ## 5.0.0-beta.30
 
 ### Patch Changes